dizanas quiver bonus
It is implemented as a visible bonus in-game, so we do the same here. Ammo can be pulled from either ammo slot and still receive the bonus, so we do not need to distinguish between the two. Weapons not using ammo slots (thrown, chinchompa, etc) do not receive the bonus.