magic damage bonus as tenths
Track magic damage bonus as an integer representing tenths of a percent. This is required to support decimal portions of magic percents, such as void magic and some of the upcoming rebalance changes.

This was already being accounted for in the calc, but needs to be shown in the UI.