updateMonster: don't alter param monster inputs
Closes #338

Setting `monster.inputs = INITIAL_MONSTER_INPUTS` was including setting `monsterCurrentHp` to 150. We handle the reset of `monsterCurrentHp` elsewhere, but the real issue is that this was reassigning the inputs field on the source object, i.e. the one from JSON.

It was not actually the second+ click that was wrong. The first call to updateMonster would edit the inputs field, then trigger the monsterCurrentHp reset hiding the misbehaviour, then subsequent clicks would not trigger that reset due to the id changing, and the dirtied inputs field would be written.