blowpipe dart as item var
This solves the export-import portion of #293, but we still need to handle warnings. This PR could be merged in the interim since it's strictly better than existing behaviour, albeit not being a full fix.

I've put this up as a separate PR from #396 in case we change the approach for errors away from that one.

If we can solve the data sourcing issue in wikisync, it will also be capable of specifying the dart id through the itemVars field.