spec calc structure (#407)
* spec calc structure

This includes the changes to the UI<->worker<->calc flow to handle special attack computation and results. It only includes the Osmumten's fang special attack as a simple example, the rest of the calc work will be in a separate PR.

Still needed: data sourcing to determine whether the current weapon has a special attack. We should disable or hide the special attack toggle on the hit distribution if inapplicable, and null out the results rows.

* spec impl

Includes:
* elder maul
* dwh
* bgs
* accursed sceptre
* arclight
* tonalztic
* dclaws
* voidwaker
* zcb
* bp
* sgs
* dhally/chally
* fang
* dds

* spec results in table and hit dist

* special attack preserve toggle state when disabled

* fix some react hook usage complaints

* adjust disabled toggle style to use opacity instead

---------

Co-authored-by: Jayden Bailey <jaydenkieran@gmail.com>