araxxor phases
Araxxor gains +35 defence during the final phase of the fight.

This is a generification of the tormented demon phase handling to support arbitrary phases on a per-monster basis. Scaling is handled as a new step in the monster scaling flow.

Previous import data is migrated using the new version metadata.