equipment select: return at least one result for matching groups (#492)
In this case, searching for `Araxyte slayer helmet (i)` was not showing any results, since the three variants are all version variants, not name variants. Therefore, all 3 were always present in the matched group, and none were shown.

This PR fixes that by showing the first result of each matching group, preferring the base variant if it is in that group. The result is that we still only see one `Araxyte slayer helmet (i)`, but it still does not appear in searches for `Slayer helmet (i)`.